Rename `retiredIn` to `retirementEpoch`.
Rename `retiredIn` to `retirementEpoch` in `PoolRetirementCertificate`.

The phrase "retired in" is somewhat ambiguous, as it might lead readers
to believe that a pool has already retired, when in fact this value can
refer to an epoch that is still in the future.

In comparison, the phrase "retirement epoch" is neutral w.r.t. whether
the epoch is in the future or the past.